Reuse existing marshaling path
Update the handler flow to mark the incoming slack interaction payload as JSON data already to keep it from being re-marshalled in the call to json.Marshal.  This allows greater reuse work common to this and existing flows (e.g. adding the prefix or printing the message to stdout before returning)